Added styles for desktop version. In index.html added anchor links to product-page. Added bestsellers block to product-page. Created separate module for loadMore functionality
