+++ app/js/pages/product-page.js
@@ -2,6 +2,7 @@ import '../modules/burgerMenu.js';
 import get from '../modules/getElement.js';
 import selectImg from '../modules/selectImg.js';
 import '../modules/changeSelectBackground.js';
+import '../modules/loadMore.js';
 
 const preview = get('.gallery__preview');
