+++ app/scss/components/_product-page.scss
@@ -1,6 +1,6 @@
 .product {
     padding-top: 20px;
-    padding-bottom: 20px;
+    padding-bottom: 40px;
 
     &__gallery {
         padding-bottom: 35px;
